+++ eva/src/js/componentsSettings.js
@@ -339,11 +339,11 @@ export default {
       option: 'titles',
       description: 'Столбцы для отображения',
       elem: 'checkbox-list',
-      items: function () {
+      items() {
         // this is modalSettings context
         return this.$store.getters.getAvailableTableTitles(
           this.idDash,
-          this.element
+          this.element,
         );
       },
       default: [],
@@ -359,10 +359,10 @@ export default {
       option: 'x',
       description: 'X axis',
       elem: 'select',
-      items: function () {
+      items() {
         return this.$store.getters.getAvailableTableTitles(
           this.idDash,
-          this.element
+          this.element,
         );
       },
     },
@@ -387,10 +387,10 @@ export default {
       option: 'y',
       description: 'Y axis',
       elem: 'select',
-      items: function () {
+      items() {
         return this.$store.getters.getAvailableTableTitles(
           this.idDash,
-          this.element
+          this.element,
         );
       },
     },
@@ -415,10 +415,10 @@ export default {
       option: 'data',
       description: 'Значение ячейки',
       elem: 'select',
-      items: function () {
+      items() {
         return this.$store.getters.getAvailableTableTitles(
           this.idDash,
-          this.element
+          this.element,
         );
       },
     },
@@ -427,10 +427,10 @@ export default {
       option: 'metadata',
       description: 'metadata',
       elem: 'select',
-      items: function () {
+      items() {
         return this.$store.getters.getAvailableTableTitles(
           this.idDash,
-          this.element
+          this.element,
         );
       },
     },
@@ -439,10 +439,10 @@ export default {
       option: 'detailValue',
       description: 'Поле для ссылки Детали',
       elem: 'select',
-      items: function () {
+      items() {
         return this.$store.getters.getAvailableTableTitles(
           this.idDash,
-          this.element
+          this.element,
         );
       },
     },
@@ -586,7 +586,7 @@ export default {
     },
     {
       relation: 'united',
-      each: function () {
+      each() {
         // each реализовано только для select и radio-group
         return this.titles && this.titles.splice
           ? [...this.titles].splice(1)
@@ -609,7 +609,7 @@ export default {
     },
     {
       relation: ['united', { axesCount: 2 }],
-      each: function () {
+      each() {
         // each реализовано только для select и radio-group, разбивает настройку для полей записи
         return this.titles && this.titles.splice
           ? [...this.titles].splice(1)